requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
openai>=1.35.0
requests>=2.32.0
python-dotenv>=1.0.1
